+++ Sources/SwiftUIKit/Extensions/Color/Color+Random.swift
@@ -10,10 +10,15 @@ import SwiftUI
 
 public extension Color {
     
-    static func random() -> Color {
+    /**
+     Generate a random color.
+     */
+    static func random(randomOpacity: Bool = false) -> Color {
         Color(
             red: .random(in: 0...1),
             green: .random(in: 0...1),
-            blue: .random(in: 0...1))
+            blue: .random(in: 0...1),
+            opacity: randomOpacity ? .random(in: 0...1) : 1
+        )
     }
 }